Moved plan runner out of init
