+++ src/lib.rs
@@ -27,5 +27,6 @@ pub mod scanner;
 pub mod expression;
 pub mod tenscript;
 pub mod init;
+pub mod plan_runner;
 
 pub use init::run;
